test: refactor file structure and improve coverage
