+++ test/unit/events/installed-app-lifecycle-event-spec.js
@@ -1,8 +1,6 @@
-const assert = require('assert').strict
 const SmartApp = require('../../../lib/smart-app')
 
-describe('uninstalled-event-spec', () => {
-	/** @type {SmartApp} */
+describe('installed-app-lifecycle-event-spec', () => {
 	let app
 	let receivedEvent
 
@@ -10,38 +8,7 @@ describe('uninstalled-event-spec', () => {
 		app = new SmartApp({logUnhandledRejections: false})
 	})
 
-	it('should handle UNINSTALL lifecycle', () => {
-		const expectedEvent = {
-			'installedApp': {
-				'installedAppId': 'd46a1c60-b6bd-4f82-b124-028e0f14a4f4',
-				'locationId': 'e1e66eab-1eab-4f09-9bb6-91da6585576d',
-				'config': {},
-				'permissions': []
-			}
-		}
-		app.uninstalled((_, event) => {
-			receivedEvent = event
-		})
-		app.handleMockCallback({
-			'lifecycle': 'UNINSTALL',
-			'executionId': 'e139dc1f-ee24-3c1a-309e-48669006817f',
-			'locale': 'en-US',
-			'version': '0.1.0',
-			'uninstallData': {
-				'installedApp': {
-					'installedAppId': 'd46a1c60-b6bd-4f82-b124-028e0f14a4f4',
-					'locationId': 'e1e66eab-1eab-4f09-9bb6-91da6585576d',
-					'config': {},
-					'permissions': []
-				}
-			},
-			'settings': {}
-		})
-
-		assert.deepStrictEqual(receivedEvent, expectedEvent)
-	})
-
-	it('should handle UNINSTALL event', () => {
+	it('should handle UNINSTALL event', async () => {
 		const expectedEvent = {
 			'installedApp': {
 				'installedAppId': '8f8004ac-789e-40a3-84db-896545a112f8',
@@ -51,7 +18,7 @@ describe('uninstalled-event-spec', () => {
 		app.uninstalled((_, event) => {
 			receivedEvent = event
 		})
-		app.handleMockCallback({
+		await app.handleMockCallback({
 			'messageType': 'EVENT',
 			'eventData': {
 				'installedApp': {
@@ -75,6 +42,6 @@ describe('uninstalled-event-spec', () => {
 			}
 		})
 
-		assert.deepStrictEqual(receivedEvent, expectedEvent)
+		expect(receivedEvent).toStrictEqual(expectedEvent)
 	})
 })
